In case of trouble make it double

+++ stories-api/src/main/java/org/craftsmenlabs/stories/api/models/config/ValidationConfig.java
@@ -39,7 +39,7 @@ public class ValidationConfig {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class ValidatorEntry {
-        private float ratingThreshold;
+        private double ratingThreshold;
         private boolean active;
 
         @Override
