+++ stories-api/src/main/java/org/craftsmenlabs/stories/api/models/items/validated/ValidatedEpic.java
@@ -17,7 +17,7 @@ import java.util.List;
 @JsonTypeName("EPIC")
 public class ValidatedEpic extends ValidatedBacklogItem<Epic> {
     @Builder
-    public ValidatedEpic(List<Violation> violations, Rating rating, Epic epic, float scoredPercentage, float missedPercentage, float scoredPoints, float missedPoints) {
+    public ValidatedEpic(List<Violation> violations, Rating rating, Epic epic, double scoredPercentage, double missedPercentage, double scoredPoints, double missedPoints) {
         super(violations, rating, epic, scoredPercentage, missedPercentage, scoredPoints, missedPoints);
     }
