+++ stories-api/src/main/java/org/craftsmenlabs/stories/api/models/items/validated/ValidatedUserStory.java
@@ -15,11 +15,11 @@ import java.util.List;
 public class ValidatedUserStory extends AbstractValidatedItem<String> {
 
     @Builder
-    public ValidatedUserStory(List<Violation> violations, Rating rating, String item, float scoredPercentage, float missedPercentage, float scoredPoints, float missedPoints) {
+    public ValidatedUserStory(List<Violation> violations, Rating rating, String item, double scoredPercentage, double missedPercentage, double scoredPoints, double missedPoints) {
         super(violations, rating, item, scoredPercentage, missedPercentage, scoredPoints, missedPoints);
     }
 
     public static ValidatedUserStory empty(){
-        return new ValidatedUserStory(Collections.emptyList(), Rating.FAIL, "", 0f, 0f, 0f, 0f);
+        return new ValidatedUserStory(Collections.emptyList(), Rating.FAIL, "", 0, 0, 0, 0);
     }
 }
