+++ stories-storyvalidator/src/main/java/org/craftsmenlabs/stories/ranking/Ranking.java
@@ -5,5 +5,5 @@ import org.craftsmenlabs.stories.api.models.items.types.BacklogItem;
 import java.util.List;
 
 public interface Ranking {
-    List<Float> getRanking(List<? extends BacklogItem> entries);
+    List<Double> getRanking(List<? extends BacklogItem> entries);
 }
